feat(alert): define a file dedicated to alerting
